+++ README.md
@@ -82,21 +82,17 @@ Conversion of enums and their usages can not be done in a single run of rector,
 that would leave the project partially converted.
 Usages seen after the enum classes have been converted will no longer be transformed.
 
-1. Configure the enum class you want to convert in [`ToNativeRector`](rector-rules.md#tonativerector)
-   in mode `ToNativeRector::USAGES` and run `vendor/bin/rector process --clear-cache` once.
+1. Configure the enum class you want to convert in [`ToNativeUsagesRector`](rector-rules.md#tonativeusagesrector)
+   and run `vendor/bin/rector process --clear-cache` once.
    Running repeatedly messes up the conversion.
-1. Set mode to `ToNativeRector::IMPLEMENTATION` and run `vendor/bin/rector process --clear-cache` once again.
+1. Switch the rector to [`ToNativeImplementationRector`](rector-rules.md#tonativeimplementationrector) 
+   and run `vendor/bin/rector process --clear-cache` once again.
 1. Review and validate the code changes for missed edge cases
    - See [Unimplemented](tests/Rector/Unimplemented)
    - `Enum::fromKey()`: Either replace usages or implement the method in your native enum.
    - `Enum::coerce()`: If only values were passed, you can replace it with `tryFrom()`.
       If keys or instances could also be passed, you might need additional logic to cover this.
    - `Enum::getDescription()`: Implement an alternative.
-
-1. TODO no longer necessary?
-   Review direct uses of the enums constant.
-   If they are used in places where enum instances are also accepted, leave them.
-   If their internal value is needed, add `->value`.
 
 ## Enum Library
