+++ rector-rules.md
@@ -1,58 +1,25 @@
-# 1 Rules Overview
+# 2 Rules Overview
 
-## ToNativeRector
+## ToNativeImplementationRector
 
 Convert usages of `BenSampo\Enum\Enum` to native PHP enums
 
 :wrench: **configure it!**
 
-- class: [`BenSampo\Enum\Rector\ToNativeRector`](src/Rector/ToNativeRector.php)
+- class: [`BenSampo\Enum\Rector\ToNativeImplementationRector`](src/Rector/ToNativeImplementationRector.php)
 
 ```php
 <?php
 
 declare(strict_types=1);
 
-use BenSampo\Enum\Rector\ToNativeRector;
+use BenSampo\Enum\Rector\ToNativeImplementationRector;
 use BenSampo\Enum\Tests\Enums\UserType;
 use Rector\Config\RectorConfig;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->ruleWithConfiguration(ToNativeRector::class, [
-        'mode' => 'usages',
-        'classes' => [
-            UserType::class,
-        ],
-    ]);
-};
-```
-
-↓
-
-```diff
--$user = UserType::ADMIN();
--$user->is(UserType::ADMIN);
-+$user = UserType::ADMIN;
-+$user === UserType::ADMIN;
-```
-
-<br>
-
-```php
-<?php
-
-declare(strict_types=1);
-
-use BenSampo\Enum\Rector\ToNativeRector;
-use BenSampo\Enum\Tests\Enums\UserType;
-use Rector\Config\RectorConfig;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->ruleWithConfiguration(ToNativeRector::class, [
-        'mode' => 'implementation',
-        'classes' => [
-            UserType::class,
-        ],
+    $rectorConfig->ruleWithConfiguration(ToNativeImplementationRector::class, [
+        UserType::class,
     ]);
 };
 ```
@@ -74,6 +41,41 @@ return static function (RectorConfig $rectorConfig): void {
 +    case ADMIN = 1;
 +    case MEMBER = 2;
  }
+```
+
+<br>
+
+## ToNativeUsagesRector
+
+Convert usages of `BenSampo\Enum\Enum` to native PHP enums
+
+:wrench: **configure it!**
+
+- class: [`BenSampo\Enum\Rector\ToNativeUsagesRector`](src/Rector/ToNativeUsagesRector.php)
+
+```php
+<?php
+
+declare(strict_types=1);
+
+use BenSampo\Enum\Rector\ToNativeUsagesRector;
+use BenSampo\Enum\Tests\Enums\UserType;
+use Rector\Config\RectorConfig;
+
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->ruleWithConfiguration(ToNativeUsagesRector::class, [
+        UserType::class,
+    ]);
+};
+```
+
+↓
+
+```diff
+-$user = UserType::ADMIN();
+-$user->is(UserType::ADMIN);
++$user = UserType::ADMIN;
++$user === UserType::ADMIN;
 ```
 
 <br>
