--- src/Rector/ToNativeRectorMode.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace BenSampo\Enum\Rector;
-
-enum ToNativeRectorMode
-{
-    case USAGES;
-    case IMPLEMENTATION;
-}
