+++ tests/Rector/implementation.php
@@ -1,10 +1,8 @@
 <?php declare(strict_types=1);
 
-use BenSampo\Enum\Rector\ToNativeRector;
+use BenSampo\Enum\Rector\ToNativeImplementationRector;
 use Rector\Config\RectorConfig;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->ruleWithConfiguration(ToNativeRector::class, [
-        'mode' => ToNativeRector::IMPLEMENTATION,
-    ]);
+    $rectorConfig->rule(ToNativeImplementationRector::class);
 };
