+++ tests/Rector/usages.php
@@ -1,10 +1,8 @@
 <?php declare(strict_types=1);
 
-use BenSampo\Enum\Rector\ToNativeRector;
+use BenSampo\Enum\Rector\ToNativeUsagesRector;
 use Rector\Config\RectorConfig;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->ruleWithConfiguration(ToNativeRector::class, [
-        'mode' => ToNativeRector::USAGES,
-    ]);
+    $rectorConfig->rule(ToNativeUsagesRector::class);
 };
